# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ultra ERP API Registry",
        "description": "Global modular API documentation for the Production-grade ERP system.",
        "contact": {
            "email": "admin@erp.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Primary API Gateway"
        }
    ],
    "paths": {
...
        "/api/v1/sales": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "List all sales",
                "operationId": "23729af480cf8e41251852845a793e7f",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "current_page": 1,
                                        "data": [
                                            {
                                                "id": 1,
                                                "sale_number": "SALE-0001",
                                                "total": 1999,
                                                "customer_name": "John Doe",
                                                "status": "completed"
                                            }
                                        ],
                                        "total": 1
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Create new sale",
                "operationId": "4547e4ebaecdac47ec1f31f4591554d8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "warehouse_id",
                                    "items"
                                ],
                                "properties": {
                                    "warehouse_id": {
                                        "type": "integer"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "number"
                                                },
                                                "unit_price": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "payments": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "account_id": {
                                                    "type": "integer"
                                                },
                                                "amount": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "customer_name": {
                                        "type": "string"
                                    },
                                    "customer_email": {
                                        "type": "string"
                                    },
                                    "currency": {
                                        "type": "string"
                                    },
                                    "branch_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sale created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Sale created successfully.",
                                    "data": {
                                        "id": 2,
                                        "sale_number": "SALE-0002",
                                        "total": 150,
                                        "customer_name": "Jane Smith"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/sales/{sale}": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Get sale details",
                "operationId": "a81e8668c17d88300ae96ffffa92a642",
                "parameters": [
                    {
                        "name": "sale",
                        "in": "path",
                        "description": "Sale ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "id": 1,
                                        "sale_number": "SALE-0001",
                                        "total": 1999,
                                        "customer_name": "John Doe",
                                        "status": "completed",
                                        "items": [
                                            {
                                                "id": 1,
                                                "product_id": 1,
                                                "quantity": 1,
                                                "selling_price": 1999
                                            }
                                        ],
                                        "payments": [
                                            {
                                                "id": 1,
                                                "account_id": 1,
                                                "amount": 1999,
                                                "method": "cash"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sale not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Sales"
                ],
                "summary": "Update sale status/notes",
                "operationId": "b53a93cadd0173bc3134d8db32287029",
                "parameters": [
                    {
                        "name": "sale",
                        "in": "path",
                        "description": "Sale ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sale updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Sale updated.",
                                    "data": {
                                        "id": 1,
                                        "status": "shipped",
                                        "notes": "Package sent via FedEx"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sale not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sales"
                ],
                "summary": "Delete sale record",
                "operationId": "808c0dfc06dd84128a14b987a50b33d5",
                "parameters": [
                    {
                        "name": "sale",
                        "in": "path",
                        "description": "Sale ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sale deleted successfully"
                    },
                    "404": {
                        "description": "Sale not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in 'Bearer {token}' format",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}